Bugfix trello 1589 incorrect stitching (#234)
* Fixed UserAgent parser + tests.

* Follow Java coding style.

* Use same code for extracting entire page size

in both ScrollPositionProvider and CssTranslatePositionProvider.

* Made `pixelRatio` in ChromeMobileEmulationDeviceSettings `double`

* Updated CHANGELOG

* Follow Java casing.

* Follow original Java casing.

* Updated constant names to match Java style.

* Fixed edge case.

+++ eyes.selenium.java/src/test/java/com/applitools/eyes/utils/ChromeMobileEmulationDeviceSettings.java
@@ -6,7 +6,7 @@ import java.util.Map;
 public class ChromeMobileEmulationDeviceSettings {
     private Map<String, Object> map;
 
-    public ChromeMobileEmulationDeviceSettings(String userAgent, int width, int height, int pixelRatio) {
+    public ChromeMobileEmulationDeviceSettings(String userAgent, int width, int height, double pixelRatio) {
         Map<String,Object> deviceMetrics = new HashMap<>();
         deviceMetrics.put("width", width);
         deviceMetrics.put("height", height);
